refactor components for JobSearch directory structure
