+++ frontend/src/App.vue
@@ -5,7 +5,7 @@
 
 <script>
 import MainNav from "@/components/Navigation/MainNav.vue";
-import TheHero from "@/components/TheHero.vue";
+import TheHero from "@/components/JobSearch/TheHero.vue";
 
 export default {
   name: "App",
