+++ frontend/src/components/JobSearch/TheHeadline.vue
