+++ frontend/src/components/JobSearch/TheHero.vue
@@ -20,8 +20,8 @@
 </template>
 
 <script>
-import TheHeadline from "@/components/TheHeadline.vue";
-import JobSearchForm from "@/components/JobSearchForm.vue";
+import TheHeadline from "@/components/JobSearch/TheHeadline.vue";
+import JobSearchForm from "@/components/JobSearch/JobSearchForm.vue";
 
 export default {
   name: "TheHero",
