+++ frontend/tests/unit/components/JobSearch/TheHeadline.test.js
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/vue";
 import { nextTick } from "vue";
-import TheHeadline from "@/components/TheHeadline.vue";
+import TheHeadline from "@/components/JobSearch/TheHeadline.vue";
 import { afterEach, beforeEach, expect, vi } from "vitest";
 
 describe("TheHeadline", () => {
